requirements.txt
pandas
requests
python-dotenv
sentence-transformers
hdbscan
scikit-learn